use user token to apply authorization for all routes

+++ src/controllers/district.ts
@@ -1,6 +1,8 @@
 import { Request, Response, NextFunction } from 'express';
 import ExpressError from '../middlewares/expressError';
 import District from '../models/district';
+import { CustomUserRequest } from '../middlewares';
+import { isAdmin, isEditor } from '../utils';
 
 export const getDistricts = async (
 	req: Request,
@@ -11,7 +13,6 @@ export const getDistricts = async (
 		const districts = await District.find();
 		res.status(200).send(districts);
 	} catch (e: any) {
-		console.error(e.message);
 		next(new ExpressError(e.message, 404));
 	}
 };
@@ -22,6 +23,10 @@ export const createDistrict = async (
 	next: NextFunction,
 ) => {
 	try {
+		const { user } = req as CustomUserRequest;
+		if (!isAdmin(user) || !isEditor(user))
+			throw new Error('YOU ARE NOT AUTHORIZED');
+
 		const { name, shippingFees } = req.body;
 		const district = new District({
 			name,
@@ -30,7 +35,6 @@ export const createDistrict = async (
 		await district.save();
 		res.status(200).send(district);
 	} catch (e: any) {
-		console.error(e.message);
 		next(new ExpressError(e.message, 404));
 	}
 };
@@ -45,7 +49,6 @@ export const getDistrict = async (
 		const district = await District.findById(district_id);
 		res.status(200).send(district);
 	} catch (e: any) {
-		console.error(e.message);
 		next(new ExpressError(e.message, 404));
 	}
 };
@@ -56,6 +59,10 @@ export const updateDistrict = async (
 	next: NextFunction,
 ) => {
 	try {
+		const { user } = req as CustomUserRequest;
+		if (!isAdmin(user) || !isEditor(user))
+			throw new Error('YOU ARE NOT AUTHORIZED');
+
 		const { district_id } = req.params;
 		const { name, shippingFees } = req.body;
 		const district = await District.findById(district_id);
@@ -64,7 +71,6 @@ export const updateDistrict = async (
 		await district.save();
 		res.status(200).send(district);
 	} catch (e: any) {
-		console.error(e.message);
 		next(new ExpressError(e.message, 404));
 	}
 };
@@ -75,11 +81,14 @@ export const deleteDistrict = async (
 	next: NextFunction,
 ) => {
 	try {
+		const { user } = req as CustomUserRequest;
+		if (!isAdmin(user) || !isEditor(user))
+			throw new Error('YOU ARE NOT AUTHORIZED');
+
 		const { district_id } = req.params;
 		await District.findByIdAndDelete(district_id);
 		res.status(200).send(district_id);
 	} catch (e: any) {
-		console.error(e.message);
 		next(new ExpressError(e.message, 404));
 	}
 };
