+++ src/controllers/profile.ts
@@ -2,7 +2,12 @@ import { Response, Request, NextFunction } from 'express';
 import ExpressError from '../middlewares/expressError';
 import Profile from '../models/profile';
 import User from '../models/user';
-import { checkPassword, genPassword } from '../utils';
+import {
+	checkPassword,
+	genPassword,
+	isAuthenticated,
+} from '../utils';
+import { CustomUserRequest } from '../middlewares';
 
 export const fetchProfile = async (
 	req: Request,
@@ -28,11 +33,19 @@ export const updateUserPhoneAndUsername = async (
 	try {
 		const { profile_id } = req.params;
 		const { phone, username } = req.body;
-
 		const profile = await Profile.findById(
 			profile_id,
 		).populate({ path: 'user' });
-		const user = await User.findById(profile?.user?._id);
+
+		const user = await User.findById(
+			(req as CustomUserRequest).user?._id,
+		);
+
+		if (
+			!isAuthenticated(user) &&
+			user?._id !== profile?.user?._id
+		)
+			throw new Error('YOU ARE NOT AUTHORIZED');
 
 		if (phone && phone.length > 0) user.phone = phone;
 		if (username && username.length > 0)
@@ -60,11 +73,20 @@ export const updateUserPassword = async (
 	try {
 		const { profile_id } = req.params;
 		const { newPassword, currentPassword } = req.body;
-
 		const profile = await Profile.findById(
 			profile_id,
 		).populate({ path: 'user' });
-		const user = await User.findById(profile?.user?._id);
+
+		const user = await User.findById(
+			(req as CustomUserRequest).user?._id,
+		);
+
+		if (
+			!isAuthenticated(user) &&
+			user?._id !== profile?.user?._id
+		)
+			throw new Error('YOU ARE NOT AUTHORIZED');
+
 		checkPassword(currentPassword, user?.password?.hash);
 
 		if (newPassword)
@@ -86,11 +108,20 @@ export const removeAccount = async (
 	try {
 		const { profile_id } = req.params;
 		const { password } = req.body;
-
 		const profile = await Profile.findById(
 			profile_id,
 		).populate({ path: 'user' });
-		const user = await User.findById(profile?.user?._id);
+
+		const user = await User.findById(
+			(req as CustomUserRequest).user?._id,
+		);
+
+		if (
+			!isAuthenticated(user) &&
+			user?._id !== profile?.user?._id
+		)
+			throw new Error('YOU ARE NOT AUTHORIZED');
+
 		checkPassword(password, user?.password?.hash);
 
 		await user?.deleteOne();
